+++ src/continuation/global_continuation_api.jl
@@ -1,4 +1,4 @@
-export global_continuation, GlobalContinuationAlgorithm, continuation_series
+export global_continuation, GlobalContinuationAlgorithm, continuation_series, stability_measures_along_continuation
 
 """
     GlobalContinuationAlgorithm
@@ -25,6 +25,9 @@ Return:
 1. `fractions_cont::Vector{Dict{Int, Float64}}`. The fractions of basins of attraction.
    `fractions_cont[i]` is a dictionary mapping attractor IDs to their basin fraction
    at the `i`-th parameter combination.
+   -  This output is different if you are using [`StabilityMeasuresAccumulator`](@ref)
+      in combination with [`AttractorSeedContinueMatch`](@ref). See the docstring
+      of [`StabilityMeasuresAccumulator`](@ref) for more details.
 2. `attractors_cont::Vector{Dict{Int, <:Any}}`. The continued attractors.
    `attractors_cont[i]` is a dictionary mapping attractor ID to the
    attractor set at the `i`-th parameter combination.
@@ -99,3 +102,4 @@ include("continuation_ascm_generic.jl")
 include("continuation_recurrences.jl")
 include("continuation_grouping.jl")
 include("aggregate_attractor_fractions.jl")
+include("continuation_stability_measures.jl")